[CORE] Added typeless fixed array
I haven't fully decided on how I want to managed typed data structures for this project yet.

but doing this typeless does incur a cost for getting and setting the data, but I expect if this becomes a serious bottleneck the approach is totally wrong.

I've also changed mem_cstd_malloc to zero out the memory under the hood.
ideally all the allocators should zero the memory despite the perf cost we should be allocating very infrequently, and if we are the liffetimes should be covered by the arena allocator