[CORE] Bitarray functions & bitset/bitarray/datastructure tests
extended bitset functions to cover bitarrays
moved bitarray/bitset/fixedarr/alloc tests out of main and into various test files.

Fixed bug/UB in bitset set range when range is the same as the word (u64) bit count i.e. set all would not set any bits as it would overflow and that is UB so many compilers return 0

I find it a little unfortunate that the inline impls for bitsets are getting huge, but I don't want to pay the cost of a function call for ops that only take a few instructions, thankfully most of these inline functions are just filled with assert which will compiled out for release