[GFX] dear c imgui (dear bindings) v1.92.2b-docking
Added default implementation of dear imgui for vulkan win32
imgui is a c++ API & impl so I've used a project called dear_bindings to generate the C interface for dear imgui.

I would usually pull the 2 projects from src, but the release for dear bindings on github doesn't build the default backends for win32 / vulkan, so as a result I generated all of the bindings & checked them in, I don't really need all of this src as I won't target OGL/SDL etc but I may throw this project into a dedicated repo for myself so I can just pull down that project, it may also be a useful reference for others on how to setup dear bindings with cmake.

There was various formatting fixes across the codebase which I fixed in this change.

I also had to expose various functions to get devices & exposed some util functions & generally marked lots of functions as static in the gfx code.