[GFX] Added gfx pool & Pool type
Pool struct is built on dynamic array with a free bit list, currently the pool isn't used but it will be used for managing task graph resources so I can pass handles around instead of pointers.

I added bitarray dynamic create/cleanup & pushed in allocInfo into the start of the dynamic bitarray, this is fine as all the bitarray functions take the head & as long as after the head we point to the backing memory all is fine, the only concern I would have is breaking alignment but that is fine because allocInfo is a ptr & the header is is also 8 byte aligned. (u64)

fixed typo in task graph assert, I was testing in releaseDebug so the test was compiled out.

eventually I will change the impl for pools to reserve all the memory but not commit but for now this is managed via the OS which is not ideal for windows but would be fine on linux.