[GFX] shuffle common vulkan vars to a global struct as func call overhead
I shuffled much of the s_gfx... vars into a g_gfx which can be externed instead of incuring the cost of a function call just to get a vulkan handle.

I will be pulling out lots of various functions into gfx_resource as i t makes sense for much of the data to be handled.

fixed Pool assert due to dynamic array current count set to 0 causing asserts for a bad index as current count is 0 even when indexing into a slot.