[CORE] Fixed bitarray used alloc tracking being too small
I was only using a bitarray 256 but the max dynamic allocations was 1024, because of this we would run out of free alloc slots when imgui is doing anything even slightly complicated, I the imgui tracking I added shows that we have aprox 350 dynamic allocs at around 0.7MiB, this numbers grows whenever new imgui panels are run, it doesn't seem doing the imgui GC cleans up enough to worry about using it.

for now in debug we can just have 4k dynamic allocs that should cover everything for a little while.