[GFX] Fixed acquire validation error & enabled synchronization2 ext
VUID-vkAcquireNextImageKHR-semaphore-01779

I've opted for calling the fence before signaling the present done semaphore just due to it being easy to implement even if it does come at a minor perf const compared to timeline semaphores.

I've upgraded to using vkQueueSubmit2KHR since I've opted to enable synchronization2 this is a core feature of 1.3. if I find I only use this for submit I will revert this to a non required as it raises the min spec for this project a bit.