[GFX] Task graph passes an additional GfxRenderContext into tasks.
we didn't have quite enough information building barriers, blitting.
I we also build the VkRenderingInfoKHR based on resources used in the render pass, we don't need to build this data for blit.

I've also moved resources into the task graph, this will need to be expanded slightly to track the last state for each resource type (potentially including external resources too)

to make this possible I added an image size to GfxImage, for most things this is usually redundant information but for renderTargets we lose the GfxAttachmentInfo i.e. format, sizetype & size