[GFX] verbose only logging for gfx_command_insert_memory_barrier
I want to ensure that the before & after code when I start using the task graph for resource transitions is the same.

in general it's probably useful to build up a interface for taking in a bit and converting it to a strig.

I opted to break the gfx_vk_image_layout_c_str() pattern to instead use the vulkan type followed by c_str as it makes using these util function much easier

it's unfortunate that I can't use a LUT as the counts are often a single bit.
I could count the first set bit & use that index as the index into a table, but as this is a debug feature I'm going to opt for the jump table over indirection table even if it is significantly slower as I will make all these functions debug only.

also added simple logging to TG so it prints the start, end & the current task name.