[GFX] imgui resize then minimize crash fix
there was a logic bug where imgui client side code can be kicked off when the renderer has not run the resize logic yet.

we could make a gfx render callback which ensures client side code runs at the correct time, but to be as simple as possible I've just exposed a test against if the surface is 0 && 0.

found a bug where we would only set the swapchain x & y size after a resize has completed but it's nice to be able to query this info instead of needing to call vkGetPhysicalDeviceSurfaceCapabilitiesKHR.