[CORE][GFX] Fixed 2 task graph resize issues
we never set s_tgArenasAllocated in cleanup, so we double free the arena memory on recompile/resize, we could follow up with a never release this until app shutdown change. probably pass in a flag to the function to tell the system if it should alloc/free.

we would leak gfx image handles in gfx_resource_image_release on every TG recompile

(I think some additional checks that all the pools are empty on shutdown would be useful too)

I added some additional checks to mem_cstd_free i.e. is the in buffer address already been freed