[CORE] Added fs funcs for getting number of files in a dir & populating a string table of file names
added win32 funcs for getting file names & counts at a dir.

renamed some funcs in the FS code so it's consistent.

some test code for the start of a image differ in the widget code

I will expand the fs code to have dirs in dir & a higher level function that builds a table of files & dirs at the target path, for now I just need to know the files in a target directory.

i'm not happy with the strlen in fs_dir_file_count
but I don't have a string / char span abstraction, so for now this will do.