[GFX] pool refactor + added more types to the pools
minor, renamed funcs for gfx pool & replaced the manual get next, get, release func calls with a macro.

updated pool create API to take in a start size & a max size instead of just max.

dropped page allocator for cstd & fixed array for dynamic.

I left a note in the memory.c code, but the issue was is the page allocator upfront allocs everything & has no way of paging in more memory for a fixed array (when the alloc is reserved but not committed) i.e. as it is all the memory is always committed to avoid this, but the just means the page allocator is worse then CSTD right now...